+++ src/darker/config.py
@@ -42,6 +42,7 @@ class DarkerConfig(TypedDict, total=False):
     skip_string_normalization: bool
     skip_magic_trailing_comma: bool
     line_length: int
+    target_version: str
     workers: int